requirements.txt
chardet==5.1.0
docx==0.2.4
cohere==4.47
Flask==2.2.3
nltk==3.8.1
numpy==1.23.2
openai==0.27.0
pandas==1.5.3
pydub==0.25.1
PyPDF2==3.0.1
python_docx==0.8.11
python_pptx==0.6.23
Requests==2.31.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
websockets==10.4
Werkzeug==2.2.3
Flask-Cors==3.0.10
Jinja2==3.1.2
httpx==0.23.3

